Working on getting chest working in VR
Added DebugTextManager.cs to help with debugging. It allows you to show values in VR instead of having to look in the console.
Added HingeJointHelper.cs. This sets some values that are needed for good hingejoint behaviour for in vr with physics hands
Added UnityEvents to VRSocket for socketing and unsocketing
Added lock and key layers so that they don't collide with each other
Added functionality for the key to be inserted into the lock