Setup xr interaction toolkit physics hands
Replaced HandPhysics script with the new NewPhysicsHand script. It is mostly the same script, but the new one removes the custom grabbing functionality.
Fixed bugs regarding the VRSocket. This script is going to be completly replaced, so I just commented the place where the error originated from.
Made a prefab for the physics hands
Setup XRGrabInteractables for 3 objects: A heavy cube, a light cube and a sphere.